refactor: handler 추상화
