+++ tomcat/src/main/java/org/apache/coyote/http11/StaticResourceHandler.java
@@ -1,12 +1,9 @@
 package org.apache.coyote.http11;
 
-import java.io.BufferedInputStream;
-import java.io.FileInputStream;
-import java.io.IOException;
 import java.net.URI;
 import java.net.URL;
 
-class StaticResourceHandler implements ResourceHandler {
+class StaticResourceHandler extends AbstractHandler {
 
     @Override
     public boolean canHandle(HttpRequest httpRequest) {
@@ -16,40 +13,10 @@ class StaticResourceHandler implements ResourceHandler {
     }
 
     @Override
-    public HttpResponse handle(HttpRequest httpRequest) {
+    public String forward(HttpRequest httpRequest) {
         URI uri = httpRequest.getUri();
         String path = uri.getPath();
-        String acceptHeader = httpRequest.header()
-                .get("Accept")
-                .orElseThrow(IllegalArgumentException::new);
-        String resourcePath = getClass().getClassLoader().getResource("static/" + path).getPath();
-        String contentType = determineContentType(resourcePath, acceptHeader);
 
-        return new HttpResponse(readStaticResource(resourcePath), contentType);
-    }
-
-    private String determineContentType(String resourcePath, String acceptHeader) {
-        String encodedContentType = "%s;charset=utf-8";
-        if (acceptHeader.startsWith("text/html") && resourcePath.endsWith(".html")) {
-            return  String.format(encodedContentType, "text/html");
-        }
-
-        if (acceptHeader.startsWith("text/css") && resourcePath.endsWith(".css")) {
-            return  String.format(encodedContentType, "text/css");
-        }
-
-        if (acceptHeader.startsWith("text/javascript") && resourcePath.endsWith(".js")) {
-            return  String.format(encodedContentType, "text/javascript");
-        }
-
-        throw new IllegalStateException();
-    }
-
-    private byte[] readStaticResource(String resourcePath) {
-        try (BufferedInputStream bufferedInputStream = new BufferedInputStream(new FileInputStream(resourcePath))) {
-            return bufferedInputStream.readAllBytes();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+        return "static/" + path;
     }
 }
